Finalization GUI inventory modificators
Finished internal logic for inventory item modification, started file saving for inventory, started logic for turn data saved

+++ Motel_management/src/main/java/model/Item.java
@@ -5,22 +5,22 @@ package model;
  */
 public class Item {
     private String name;
-    private int price;
-    private int quantity;
-    private int itemID;
+    private long price;
+    private long quantity;
+    private long itemID;
     
-    public Item(String name, int value, int quantity, int itemID){
+    public Item(String name, long value, long quantity, long itemID){
         this.itemID = itemID;
         this.name = name;
         this.price = value;
         this.quantity = quantity;
     }
     
-    public void itemSold(int quantitySold){
+    public void itemSold(long quantitySold){
         quantity = quantity - quantitySold;
     }
     
-    public void itemAdded(int quantityAdded){
+    public void itemAdded(long quantityAdded){
         quantity  = quantity + quantityAdded;
     }
 
@@ -35,28 +35,28 @@ public class Item {
     /**
      * @return the price
      */
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
 
     /**
      * @return the quantity
      */
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
 
     /**
      * @param quantity the quantity to set
      */
-    public void setQuantity(int quantity) {
+    public void setQuantity(long quantity) {
         this.quantity = quantity;
     }
 
     /**
      * @return the itemID
      */
-    public int getItemID() {
+    public long getItemID() {
         return itemID;
     }
